refactored and retraited indexer, refactored validator
