+++ zingo_test_vectors/src/lib.rs
@@ -41,6 +41,9 @@ pub const REG_T_ADDR_FROM_ABANDONART: &str = "tmBsTi2xWTjUdEXnuTceL7fecEQKeWaPDJ
 /// TODO: Add Doc Comment Here!
 pub const FUND_OFFLOAD_ORCHARD_ONLY: &str = "uregtest1duh3glf8uk5he5cpmlzsfvkn34de4uudyahdr7p6j0p6zs2tujgdxqmzgvtquwc5cphwufku93a0p5ksxzwx0qk92kkd5nrdzs5tngw6";
 
+/// Zebrad default miner address. Regtest/Testnet transparent address for [Abandon Abandon .. Art] seed (entropy all zeros)
+pub const ZEBRAD_DEFAULT_MINER: &str = "tmBsTi2xWTjUdEXnuTceL7fecEQKeWaPDJd";
+
 /// Regtest block rewards
 pub mod block_rewards {
     /// TODO: Add Doc Comment Here!
